+++ scripts/pull_and_deploy.sh
@@ -2,13 +2,13 @@
 
 COMPOSE_FILE="${1:-docker-compose.yml}"
 
-echo "Using ${COMPOSE_FILE}" 
+echo "Using ${COMPOSE_FILE}"
 
 git pull && \
 docker compose -f $COMPOSE_FILE pull && \
 docker compose -f $COMPOSE_FILE build && \
 #docker compose -f $COMPOSE_FILE up -d && \
 echo "Done!"
-#sleep 5 
+#sleep 5
 #&& \
 #docker compose logs