+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 Adding notes to what is being changed from time to time
 
-## [0.2.0] 03.04.2023
+## [0.2.6] 03.04.2023
 ### Added
 - performance test script that generates stats for each file
 - option to load and download ro_legal_fl model
@@ -9,6 +9,7 @@ Adding notes to what is being changed from time to time
 - pipeline fixes
 - stop auto-deploy of API service
 - performance tests with docker-compose.yml
+- pre-commit hooks
 
 
 ## [0.1.1] 22.03.2023
@@ -40,4 +41,4 @@ Adding notes to what is being changed from time to time
 - prepare Dockerfile commands to install rust
 ### Changed
 - text is being returned from blocks instead of raw text
-- cleaned worker 
+- cleaned worker